html/semantics/embedded-content/the-iframe-element/iframe-loading-lazy-nav-link-click.html

<!DOCTYPE html>
<title>Navigating iframe loading='lazy' before it is loaded: link click</title>
<iframe name="iframe" src="support/blank.htm?src" loading="lazy" onload="window.gotIframeLoadEvent = true;"></iframe>
<a href="support/blank.htm?nav" target="iframe"></a>
<script>
const iframe = document.querySelector('iframe');
const iframeLoaded = new Promise(resolve => {
  window.gotIframeLoadEvent ? resolve() : iframe.onload = resolve;
});
const a = document.querySelector('a');
a.click();
</script>
<!-- Loading testharness.js here is intentional to reproduce a bug in WebKit. -->
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
setup({single_test: true});
iframeLoaded.then(() => {
  step_timeout(() => {
    assert_equals(iframe.contentWindow.location.href, new URL("support/blank.htm?nav", location.href).href);
    done();
  }, 1000);
});
</script>
